+++ resources/user.py
@@ -1,7 +1,7 @@
 from flask_restful import Resource, reqparse
 from models.user import UserModel
 from werkzeug.security import check_password_hash
-from flask_jwt_extended import create_access_token, create_refresh_token
+from flask_jwt_extended import create_access_token, create_refresh_token, jwt_required, get_jwt
 
 user_parser = reqparse.RequestParser()
 user_parser.add_argument("username",
@@ -26,7 +26,7 @@ class UserRegister(Resource):
         return {"message": "User Created Successfully"}, 201
 
 class User(Resource):
-    @classmethod    #Since we dont need to access object of this class we can make classmethod
+    @classmethod    #Since we dont need object of this class we can make classmethod
     def get(cls,user_id):
         user = UserModel.find_by_userid(user_id)
         if user:
@@ -34,7 +34,12 @@ class User(Resource):
         return {"message": "User Not Found"},404
 
     @classmethod
+    @jwt_required()
     def delete(cls,user_id):
+        claims = get_jwt()  #jwt claims details
+        if not claims["is_admin"]:
+            #if user claims to be admin then only, he can performs delete operation
+            return{"message": "Admin privileges required for this operation"},401
         user = UserModel.find_by_userid(user_id)
         if user:
             user.delete_from_db()
@@ -46,6 +51,7 @@ class UserLogin(Resource):
         data = user_parser.parse_args()
         user = UserModel.find_by_username(data["username"])
         if user and check_password_hash(user.password, data["password"]):
+            #if user credentials are correct generate access and refresh token
             access_token = create_access_token(identity=user.id, fresh=True)
             refresh_token = create_refresh_token(user.id)
             return {
